Add new SVD file from MCUXpresso SDL Builder
This file is available from NXP. You need to build and download an
MCUXpresso toolchain[1] for the i.MX 1052 or its development board. The
toolchain contains the SVD file.

I haven't reviewed the new file or compared it to the old one. The diff
is huge. My reason for wanting to use the new file is that the old one
is missing important GPIO registers (DR_SET, DR_CLEAR, DR_TOGGLE).

[1] https://mcuxpresso.nxp.com/en/welcome